Lead Converted to Application

+++ app/Http/Repositories/ApplicationRepository.php
@@ -4,6 +4,7 @@ namespace App\Http\Repositories;
 use App\Http\Contracts\RepositoryContract;
 use App\Models\Application;
 use Exception;
+use Illuminate\Support\Facades\Gate;
 
 class ApplicationRepository implements RepositoryContract
 {
@@ -13,6 +14,15 @@ class ApplicationRepository implements RepositoryContract
             $offset    = request()->input('offset') ?? 10;
             $relations = [];
             $fields    = ['id', 'lead_id', 'user_id', 'name', 'status'];
+
+            if (! Gate::inspect('view-any', Application::class)->allowed()) {
+                return messageResponse("Sorry, You can't access this resource", 403);
+            }
+
+            $userId = auth()->guard('api')->check() ?
+            auth()->guard('api')->user()->role === 'COUNSELOR'
+            ? auth()->guard('api')->user()->id : null
+            : null;
 
             if (request()->has('fields') && request()->input('fields')) {
                 $fields = gettype(request()->input('fields')) === 'array' ? request()->input('fields') : explode(',', request()->input('fields'));
@@ -26,11 +36,9 @@ class ApplicationRepository implements RepositoryContract
                 ->with($relations)
                 ->select($fields)
                 ->when(request()->input('lead_id'), fn($q) => $q->where('lead_id', request()->input('lead_id')))
-                ->when(request()->input('user_id'), fn($q) => $q->where('user_id', request()->input('user_id')))
+                ->when($userId, fn($q) => $q->where('user_id', $userId))
                 ->when(request()->input('status'), fn($q) => $q->where('status', request()->input('status')))
-                ->when(request()->input('search'), function ($q) {
-                    return $q->$q->where('name', 'like', '%' . request()->input('search') . '%');
-                });
+                ->when(request()->input('search'), fn($q) => $q->where('name', 'like', '%' . request()->input('search') . '%'));
 
             if (request()->has('get_all') && (int) request()->input('get_all') === 1) {
                 return entityResponse($docs->get());
@@ -45,6 +53,7 @@ class ApplicationRepository implements RepositoryContract
     public function get($id)
     {
         try {
+            $condition = ['id' => $id];
             $relations = [];
             $fields    = ['id', 'lead_id', 'user_id', 'name', 'description', 'status'];
 
@@ -56,8 +65,12 @@ class ApplicationRepository implements RepositoryContract
                 $relations = gettype(request()->input('relations')) === 'array' ? request()->input('relations') : explode(',', request()->input('relations'));
             }
 
-            if (! $doc = Application::query()->with($relations)->select($fields)->first()) {
+            if (! $doc = Application::query()->with($relations)->select($fields)->where($condition)->first()) {
                 return messageResponse();
+            }
+
+            if (! Gate::inspect('view', $doc)->allowed()) {
+                return messageResponse("Sorry, You can't access this application", 403);
             }
 
             return entityResponse($doc);
@@ -69,6 +82,10 @@ class ApplicationRepository implements RepositoryContract
     public function store($request)
     {
         try {
+            if (! Gate::inspect('create', Application::class)->allowed()) {
+                return messageResponse("Sorry, You can't access this resource", 403);
+            }
+
             Application::query()->create($request->validated());
             return messageResponse('Application added successfully', 201, 'success');
         } catch (Exception $e) {
@@ -83,6 +100,10 @@ class ApplicationRepository implements RepositoryContract
                 return messageResponse();
             }
 
+            if (! Gate::inspect('update', $doc)->allowed()) {
+                return messageResponse("Sorry, You can't access this application", 403);
+            }
+
             $doc->update($request->validated());
             return messageResponse('Application updated successfully', 200, 'success');
         } catch (Exception $e) {
@@ -95,6 +116,10 @@ class ApplicationRepository implements RepositoryContract
         try {
             if (! $doc = Application::query()->where(['id' => $id])->first()) {
                 return messageResponse();
+            }
+
+            if (! Gate::inspect('delete', $doc)->allowed()) {
+                return messageResponse("Sorry, You can't access this application", 403);
             }
 
             $doc->delete();
