+++ app/Http/Repositories/LeadRepository.php
@@ -4,6 +4,7 @@ namespace App\Http\Repositories;
 use App\Http\Contracts\RepositoryContract;
 use App\Models\Lead;
 use Exception;
+use Illuminate\Support\Facades\Gate;
 
 class LeadRepository implements RepositoryContract
 {
@@ -13,6 +14,12 @@ class LeadRepository implements RepositoryContract
             $offset    = request()->input('offset') ?? 10;
             $relations = [];
             $fields    = ['id', 'type', 'name', 'email', 'phone', 'status'];
+
+            if (! Gate::inspect('view-any', Lead::class)->allowed()) {
+                return messageResponse("Sorry, You can't access this resource", 403);
+            }
+
+            $userRole = auth()->guard('api')->check() ? auth()->guard('api')->user()->role : null;
 
             if (request()->has('fields') && request()->input('fields')) {
                 $fields = gettype(request()->input('fields')) === 'array' ? request()->input('fields') : explode(',', request()->input('fields'));
@@ -29,7 +36,8 @@ class LeadRepository implements RepositoryContract
                 ->when(request()->input('status'), fn($q) => $q->where('status', request()->input('status')))
                 ->when(request()->input('search'), function ($q) {
                     return $q->where(fn() => $q->where('name', 'like', '%' . request()->input('search') . '%')->orWhere('email', 'like', '%' . request()->input('search') . '%'));
-                });
+                })
+                ->when($userRole === 'COUNSELOR', fn($q) => $q->whereHas('maintainers', fn($query) => $query->where('user_id', auth()->guard('api')->id())));
 
             if (request()->has('get_all') && (int) request()->input('get_all') === 1) {
                 return entityResponse($docs->get());
@@ -44,6 +52,7 @@ class LeadRepository implements RepositoryContract
     public function get($id)
     {
         try {
+            $condition = ['id' => $id];
             $relations = [];
             $fields    = ['id', 'type', 'name', 'email', 'phone', 'address', 'status'];
 
@@ -55,8 +64,12 @@ class LeadRepository implements RepositoryContract
                 $relations = gettype(request()->input('relations')) === 'array' ? request()->input('relations') : explode(',', request()->input('relations'));
             }
 
-            if (! $doc = Lead::query()->with($relations)->select($fields)->first()) {
+            if (! $doc = Lead::query()->with($relations)->select($fields)->where($condition)->first()) {
                 return messageResponse();
+            }
+
+            if (! Gate::inspect('view', $doc)->allowed()) {
+                return messageResponse("Sorry, You can't access this lead", 403);
             }
 
             return entityResponse($doc);
@@ -68,6 +81,10 @@ class LeadRepository implements RepositoryContract
     public function store($request)
     {
         try {
+            if (! Gate::inspect('create', Lead::class)->allowed()) {
+                return messageResponse("Sorry, You can't access this lead", 403);
+            }
+
             Lead::query()->create($request->validated());
             return messageResponse('Lead added successfully', 201, 'success');
         } catch (Exception $e) {
@@ -82,6 +99,10 @@ class LeadRepository implements RepositoryContract
                 return messageResponse();
             }
 
+            if (! Gate::inspect('update', $doc)->allowed()) {
+                return messageResponse("Sorry, You can't access this lead", 403);
+            }
+
             $doc->update($request->validated());
             return messageResponse('Lead updated successfully', 200, 'success');
         } catch (Exception $e) {
@@ -94,6 +115,10 @@ class LeadRepository implements RepositoryContract
         try {
             if (! $doc = Lead::query()->where(['id' => $id])->first()) {
                 return messageResponse();
+            }
+
+            if (! Gate::inspect('delete', $doc)->allowed()) {
+                return messageResponse("Sorry, You can't access this lead", 403);
             }
 
             $doc->delete();
